requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
psycopg2-binary==2.9.2
pycparser==2.21
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
djangorestframework~=3.12.4